+++ src/gui/control/listbox.h
@@ -45,7 +45,7 @@ public:
     bool handleMouseWheel(int x, int y) override;
     bool handleClick(int x, int y) override;
     void initGL() override;
-    void render(const glm::vec2 &offset, const std::string &textOverride) const override;
+    void render(const glm::ivec2 &offset, const std::string &textOverride) const override;
     void stretch(float x, float y) override;
 
     void setFocus(bool focus) override;
